Moved files around, changed header file locations, added app style format
